moved non-async helper classes out of async folder
